+++ src/parser/ty/path_type.rs
@@ -1,5 +1,7 @@
+use core::fmt;
+
 use crate::{
-    ast::{Identifier, PathRoot, PathType, SelfType},
+    ast::{Identifier, PathExpr, PathPatt, PathRoot, PathType, SelfType},
     error::ErrorsEmitted,
     token::Token,
 };
@@ -96,5 +98,88 @@ impl PathType {
         ////////////////////////////////////////////////////////////////////////////////
 
         Ok(path_type)
+    }
+}
+
+impl From<Identifier> for PathType {
+    fn from(value: Identifier) -> Self {
+        PathType {
+            associated_type_path_prefix_opt: None,
+            type_name: value,
+        }
+    }
+}
+
+impl From<PathExpr> for PathType {
+    fn from(value: PathExpr) -> Self {
+        let mut path_segment_names: Vec<Identifier> = Vec::new();
+
+        path_segment_names.push(Identifier::from(&value.path_root.to_string()));
+
+        match value.tree_opt {
+            Some(v) => {
+                v.into_iter().for_each(|i| path_segment_names.push(i));
+            }
+            None => (),
+        }
+
+        let type_name = path_segment_names.pop().expect("empty path expression");
+
+        PathType {
+            associated_type_path_prefix_opt: {
+                if path_segment_names.is_empty() {
+                    None
+                } else {
+                    Some(path_segment_names)
+                }
+            },
+            type_name,
+        }
+    }
+}
+
+impl From<PathPatt> for PathType {
+    fn from(value: PathPatt) -> Self {
+        let mut path_segment_names: Vec<Identifier> = Vec::new();
+
+        path_segment_names.push(Identifier::from(&value.path_root.to_string()));
+
+        match value.tree_opt {
+            Some(v) => {
+                v.into_iter().for_each(|i| path_segment_names.push(i));
+            }
+            None => (),
+        }
+
+        let type_name = path_segment_names.pop().expect("empty path expression");
+
+        PathType {
+            associated_type_path_prefix_opt: {
+                if path_segment_names.is_empty() {
+                    None
+                } else {
+                    Some(path_segment_names)
+                }
+            },
+            type_name,
+        }
+    }
+}
+
+impl fmt::Display for PathType {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        let mut segments: Vec<String> = Vec::new();
+
+        if let Some(v) = &self.associated_type_path_prefix_opt {
+            for i in v {
+                segments.push(i.to_string());
+            }
+        }
+
+        segments.push(self.type_name.to_string());
+
+        let full_path = segments.join("::");
+
+        write!(f, "{}", full_path)
     }
 }
